requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
line-bot-sdk==3.5.0
python-dotenv==1.0.0
httpx==0.26.0
pydantic==2.5.3